+++ src/pytraits/targets/resolutions.py
@@ -16,11 +16,7 @@
    limitations under the License.
 '''
 
-from pytraits.core import Singleton, type_safe, type_converted
-from pytraits.combiner import combine_class
-from pytraits.extendable import extendable
-from pytraits.trait_composer import add_traits
 
-__version__ = "0.15.0"
-__all__ = ["Singleton", "combine_class", "extendable", "add_traits",
-           "type_safe", "type_converted"]
+class Resolutions:
+    def __init__(self, resolutions):
+        self.__resolutions = resolutions